chore: rename the authentication helper
