+++ src/app/utils/authentication-check.ts
